+++ CHANGELOG.md
@@ -52,6 +52,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Improvements
 
 * [#595](https://github.com/provenance-io/cosmos-sdk/pull/595) Provenance: Create SimulateFromSeedProv which is the same as SimulateFromSeed but also returns the simulation's end block time which is needed for some sims tests.
+* [#604](https://github.com/provenance-io/cosmos-sdk/pull/604) Provenance: Create bankkeeper.MigrateParamsProv so we can finish the v3-v4 bank module migration.
 
 ### Bug Fixes